Move lifespan tests out of obsolete middleware test module
